# src/domain/evaluation/dto/create-evaluation.dto.ts
import { IsString } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class CreateEvaluationDto {
	@ApiProperty({ example: 'Initial Assessment' })
	@IsString()
	title!: string;
}