# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Get hello message","tags":["app"]}},"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"summary":"Create a new user","tags":["user"]},"get":{"operationId":"UserController_findAll","parameters":[{"name":"filter","required":false,"in":"query","description":"JSON object containing filter criteria","schema":{"$ref":"#/components/schemas/Object"}},{"name":"page","required":false,"in":"query","description":"Page number for pagination","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"example":10,"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Retrieve all users","tags":["user"]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Retrieve a single user by ID","tags":["user"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"summary":"Update an existing user by ID","tags":["user"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Delete a user by ID","tags":["user"]}},"/enterprise":{"post":{"operationId":"EnterpriseController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEnterpriseDto"}}}},"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"summary":"Create a new enterprise","tags":["enterprise"]},"get":{"operationId":"EnterpriseController_findAll","parameters":[{"name":"filter","required":false,"in":"query","description":"JSON object containing filter criteria","schema":{"$ref":"#/components/schemas/Object"}},{"name":"page","required":false,"in":"query","description":"Page number for pagination","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"example":10,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"List all enterprises","tags":["enterprise"]}},"/enterprise/{id}":{"get":{"operationId":"EnterpriseController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Get enterprise by ID","tags":["enterprise"]},"patch":{"operationId":"EnterpriseController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEnterpriseDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Update enterprise by ID","tags":["enterprise"]},"delete":{"operationId":"EnterpriseController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Delete enterprise by ID","tags":["enterprise"]}},"/student":{"post":{"operationId":"StudentController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStudentDto"}}}},"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"summary":"Create a new student","tags":["student"]},"get":{"operationId":"StudentController_findAll","parameters":[{"name":"filter","required":false,"in":"query","description":"JSON object containing filter criteria","schema":{"$ref":"#/components/schemas/Object"}},{"name":"page","required":false,"in":"query","description":"Page number for pagination","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"example":10,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"List all students","tags":["student"]}},"/student/{id}":{"get":{"operationId":"StudentController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Get student by ID","tags":["student"]},"patch":{"operationId":"StudentController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStudentDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Update student by ID","tags":["student"]},"delete":{"operationId":"StudentController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Delete student by ID","tags":["student"]}},"/monitoring":{"post":{"operationId":"MonitoringController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMonitoringDto"}}}},"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"summary":"Create a new monitoring record","tags":["monitoring"]},"get":{"operationId":"MonitoringController_findAll","parameters":[{"name":"filter","required":false,"in":"query","description":"JSON object containing filter criteria","schema":{"$ref":"#/components/schemas/Object"}},{"name":"page","required":false,"in":"query","description":"Page number for pagination","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"example":10,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"List all monitoring records","tags":["monitoring"]}},"/monitoring/{id}":{"get":{"operationId":"MonitoringController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Get monitoring record by ID","tags":["monitoring"]},"patch":{"operationId":"MonitoringController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMonitoringDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Update monitoring record by ID","tags":["monitoring"]},"delete":{"operationId":"MonitoringController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Delete monitoring record by ID","tags":["monitoring"]}},"/evaluation":{"post":{"operationId":"EvaluationController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEvaluationDto"}}}},"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"summary":"Create a new evaluation","tags":["evaluation"]},"get":{"operationId":"EvaluationController_findAll","parameters":[{"name":"filter","required":false,"in":"query","description":"JSON object containing filter criteria","schema":{"$ref":"#/components/schemas/Object"}},{"name":"page","required":false,"in":"query","description":"Page number for pagination","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"example":10,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"List all evaluations","tags":["evaluation"]}},"/evaluation/{id}":{"get":{"operationId":"EvaluationController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Get evaluation by ID","tags":["evaluation"]},"patch":{"operationId":"EvaluationController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEvaluationDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Update evaluation by ID","tags":["evaluation"]},"delete":{"operationId":"EvaluationController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Delete evaluation by ID","tags":["evaluation"]}},"/evaluation-field-type":{"post":{"operationId":"EvaluationFieldTypeController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEvaluationFieldTypeDto"}}}},"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"summary":"Create a new evaluation field type","tags":["evaluation-field-type"]},"get":{"operationId":"EvaluationFieldTypeController_findAll","parameters":[{"name":"filter","required":false,"in":"query","description":"JSON object containing filter criteria","schema":{"$ref":"#/components/schemas/Object"}},{"name":"page","required":false,"in":"query","description":"Page number for pagination","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"example":10,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"List all evaluation field types","tags":["evaluation-field-type"]}},"/evaluation-field-type/{id}":{"get":{"operationId":"EvaluationFieldTypeController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Get evaluation field type by ID","tags":["evaluation-field-type"]},"patch":{"operationId":"EvaluationFieldTypeController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEvaluationFieldTypeDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Update evaluation field type by ID","tags":["evaluation-field-type"]},"delete":{"operationId":"EvaluationFieldTypeController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Delete evaluation field type by ID","tags":["evaluation-field-type"]}},"/evaluation-question":{"post":{"operationId":"EvaluationQuestionController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEvaluationQuestionDto"}}}},"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"summary":"Create a new evaluation question","tags":["evaluation-question"]},"get":{"operationId":"EvaluationQuestionController_findAll","parameters":[{"name":"filter","required":false,"in":"query","description":"JSON object containing filter criteria","schema":{"$ref":"#/components/schemas/Object"}},{"name":"page","required":false,"in":"query","description":"Page number for pagination","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"example":10,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"List all evaluation questions","tags":["evaluation-question"]}},"/evaluation-question/{id}":{"get":{"operationId":"EvaluationQuestionController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Get evaluation question by ID","tags":["evaluation-question"]},"patch":{"operationId":"EvaluationQuestionController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEvaluationQuestionDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Update evaluation question by ID","tags":["evaluation-question"]},"delete":{"operationId":"EvaluationQuestionController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Delete evaluation question by ID","tags":["evaluation-question"]}},"/evaluation-field":{"post":{"operationId":"EvaluationFieldController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEvaluationFieldDto"}}}},"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"summary":"Create a new evaluation field","tags":["evaluation-field"]},"get":{"operationId":"EvaluationFieldController_findAll","parameters":[{"name":"filter","required":false,"in":"query","description":"JSON object containing filter criteria","schema":{"$ref":"#/components/schemas/Object"}},{"name":"page","required":false,"in":"query","description":"Page number for pagination","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"example":10,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"List all evaluation fields","tags":["evaluation-field"]}},"/evaluation-field/{id}":{"get":{"operationId":"EvaluationFieldController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Get evaluation field by ID","tags":["evaluation-field"]},"patch":{"operationId":"EvaluationFieldController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEvaluationFieldDto"}}}},"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Update evaluation field by ID","tags":["evaluation-field"]},"delete":{"operationId":"EvaluationFieldController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Delete evaluation field by ID","tags":["evaluation-field"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"201":{"description":""}},"summary":"Authenticate user and receive JWT token","tags":["auth"]}}},"info":{"title":"Instituto Diomício Freitas API","description":"API documentation for the backend","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","example":"john_doe"},"password":{"type":"string","example":"P@ssword123"},"email":{"type":"string","example":"user@example.com"}},"required":["username","password","email"]},"Object":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"CreateEnterpriseDto":{"type":"object","properties":{"name":{"type":"string","example":"Acme Ltda"},"cnpj":{"type":"string","example":"12345678000199"},"phone":{"type":"string","example":"+551130000000"}},"required":["name","cnpj","phone"]},"UpdateEnterpriseDto":{"type":"object","properties":{}},"CreateStudentDto":{"type":"object","properties":{"name":{"type":"string","example":"Jane Doe"},"phone":{"type":"string","example":"+5511999998888"},"date_birthday":{"type":"string","example":"2000-01-01"},"cpf":{"type":"string","example":"12345678901"},"enterpriseId":{"type":"number","example":1}},"required":["name","phone","date_birthday","cpf"]},"UpdateStudentDto":{"type":"object","properties":{}},"CreateMonitoringDto":{"type":"object","properties":{"notes":{"type":"string","example":"Estudante demonstra melhorias na comunicação"},"studentId":{"type":"number","example":5},"evaluationId":{"type":"number","example":2}},"required":["studentId","evaluationId"]},"UpdateMonitoringDto":{"type":"object","properties":{}},"CreateEvaluationDto":{"type":"object","properties":{"title":{"type":"string","example":"Initial Assessment"}},"required":["title"]},"UpdateEvaluationDto":{"type":"object","properties":{}},"CreateEvaluationFieldTypeDto":{"type":"object","properties":{"description":{"type":"string","example":"Text input"}}},"UpdateEvaluationFieldTypeDto":{"type":"object","properties":{}},"CreateEvaluationQuestionDto":{"type":"object","properties":{"question":{"type":"string","example":"Quão satisfeito você está com o atendimento?"},"evaluationId":{"type":"number","example":3}},"required":["question","evaluationId"]},"UpdateEvaluationQuestionDto":{"type":"object","properties":{}},"CreateEvaluationFieldDto":{"type":"object","properties":{"label":{"type":"string","example":"Nível de comunicação"},"questionId":{"type":"number","example":1},"typeId":{"type":"number","example":2}},"required":["label","questionId","typeId"]},"UpdateEvaluationFieldDto":{"type":"object","properties":{}},"SignInDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"P@ssword123"}},"required":["email","password"]}}}}